refactor: move Task-related classes to tasks package
