+++ src/main/java/module-info.java
@@ -7,4 +7,8 @@ module uni.usic.taskmanagerv1javafx {
             
     opens uni.usic.taskmanager to javafx.fxml;
     exports uni.usic.taskmanager;
+    exports uni.usic.taskmanager.views.tasks;
+    opens uni.usic.taskmanager.views.tasks to javafx.fxml;
+    exports uni.usic.taskmanager.controller.tasks;
+    opens uni.usic.taskmanager.controller.tasks to javafx.fxml;
 }
