--- src/main/java/uni/usic/domain/entity/Viewable.java
@@ -1,5 +1,0 @@
-package uni.usic.domain.entity;
-
-public interface Viewable {
-    void viewable();
-}
