--- src/main/java/uni/usic/domain/entity/maintasks/StudyTask.java
@@ -1,19 +1,0 @@
-package uni.usic.domain.entity.maintasks;
-
-import uni.usic.domain.enums.TaskPriority;
-
-import java.time.LocalDate;
-
-public class StudyTask extends Task {
-    private String subject;
-    private String studyType;
-    private int estimateTime;
-
-    public StudyTask(String id, String title, String description, LocalDate startDate, LocalDate endDate, TaskPriority priority) {
-        super(id, title, description, startDate, endDate, priority);
-    }
-
-    public void startStudying() {
-
-    }
-}
