+++ src/main/java/uni/usic/domain/entity/tasks/Viewable.java
@@ -1,0 +1,5 @@
+package uni.usic.domain.entity.tasks;
+
+public interface Viewable {
+    void viewable();
+}
