+++ src/main/java/uni/usic/domain/entity/tasks/enums/TaskPriority.java
@@ -1,0 +1,5 @@
+package uni.usic.domain.entity.tasks.enums;
+
+public enum TaskPriority {
+    HIGH, MEDIUM, LOW;
+}
