+++ src/main/java/uni/usic/domain/entity/tasks/enums/TaskProgress.java
@@ -1,0 +1,5 @@
+package uni.usic.domain.entity.tasks.enums;
+
+public enum TaskProgress {
+    TO_DO, IN_PROGRESS, DONE
+}
