--- src/main/java/uni/usic/domain/enums/TaskPriority.java
@@ -1,5 +1,0 @@
-package uni.usic.domain.enums;
-
-public enum TaskPriority {
-    HIGH, MEDIUM, LOW;
-}
