--- src/main/java/uni/usic/domain/enums/TaskProgress.java
@@ -1,5 +1,0 @@
-package uni.usic.domain.enums;
-
-public enum TaskProgress {
-    TO_DO, IN_PROGRESS, DONE
-}
