+++ src/main/java/uni/usic/taskmanager/TaskApplication.java
@@ -13,6 +13,7 @@ import javafx.scene.layout.VBox;
 import javafx.scene.text.Font;
 import javafx.scene.text.FontWeight;
 import javafx.stage.Stage;
+import uni.usic.taskmanager.views.tasks.TaskList;
 
 import java.io.IOException;
