--- src/main/java/uni/usic/taskmanager/TaskController.java
@@ -1,14 +1,0 @@
-package uni.usic.taskmanager;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class TaskController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onHelloButtonClick() {
-        welcomeText.setText("Welcome to JavaFX Application!");
-    }
-}
