+++ src/main/java/uni/usic/taskmanager/controller/tasks/TaskController.java
@@ -1,0 +1,14 @@
+package uni.usic.taskmanager.controller.tasks;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+public class TaskController {
+    @FXML
+    private Label welcomeText;
+
+    @FXML
+    protected void onHelloButtonClick() {
+        welcomeText.setText("Welcome to JavaFX Application!");
+    }
+}
